Rename nomenclature use service, form & query generators
The proper name of these patters comes with the prefix , but at a
programmer convinience point of view using just terms like service, form and query
makes the CLI more convinient.

Also giveing the class name with the prefix Service, Query and Form. And using terms like ThingService, ThingQuery and ThingForm would improve the redability of our code. Thus I propose these changes in nomenclature.
