+++ app/services/thing.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class ThingService
+  def initialize; end
+
+  def run
+    # Write service logic here
+  end
+end
