+++ lib/generators/form/form_generator.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-class FormObjectGenerator < Rails::Generators::NamedBase
-  source_root File.expand_path('../templates', __FILE__)
+class FormGenerator < Rails::Generators::NamedBase
+  source_root File.expand_path('templates', __dir__)
   argument :attributes, type: :array, default: [], banner: 'field field'
 
-  def copy_form_object_file
-    template 'form_object.rb.tt', File.join('app', 'forms', class_path, "#{file_name}_form.rb")
+  def copy_form_file
+    template 'form.rb.tt', File.join('app', 'forms', class_path, "#{file_name}_form.rb")
   end
 
   hook_for :test_framework
