+++ lib/generators/form/templates/form.rb.tt
@@ -1,7 +1,7 @@
 <% module_namespacing do -%>
 class <%= class_name %>Form
-  include ActiveModel::Model 
-  
+  include ActiveModel::Model
+
   attr_accessor <%= attribute_symbols.join(", ") %>
 
   def self.model_name
