+++ lib/generators/query/templates/query.rb.tt
@@ -1,6 +1,6 @@
 <% module_namespacing do -%>
 class <%= class_name %>Query
-  attr_reader :relation  
+  attr_reader :relation
   def initialize(relation = <%= class_name %>.all)
     @relation = relation
   end
