+++ lib/generators/rspec/query/templates/query_spec.rb.tt
