+++ lib/generators/rspec/service/service_generator.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
 module Rspec
-  class ServiceObjectGenerator < Rails::Generators::NamedBase
-    source_root File.expand_path('../templates', __FILE__)
+  class ServiceGenerator < Rails::Generators::NamedBase
+    source_root File.expand_path('templates', __dir__)
 
     argument :attributes, type: :array, default: [], banner: 'field field'
 
-    def copy_service_object_file
-      template 'service_object_spec.rb.tt', File.join('spec', 'services', class_path, "#{file_name}_spec.rb")
+    def copy_service_file
+      template 'service_spec.rb.tt', File.join('spec', 'services', class_path, "#{file_name}_service_spec.rb")
     end
 
     private
