+++ lib/generators/rspec/service/templates/service_spec.rb.tt
@@ -1,6 +1,6 @@
 require 'rails_helper'
 <% module_namespacing do -%>
-  RSpec.describe <%= class_name %> do
+  RSpec.describe <%= class_name %>Service do
     pending "add some tests to the <%= file_name %>_spec.rb"
   end
 <% end -%>
