+++ lib/generators/service/service_generator.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-class ServiceObjectGenerator < Rails::Generators::NamedBase
-  source_root File.expand_path('../templates', __FILE__)
+class ServiceGenerator < Rails::Generators::NamedBase
+  source_root File.expand_path('templates', __dir__)
 
   argument :attributes, type: :array, default: [], banner: 'field field'
 
-  def copy_service_object_file
-    template 'service_object.rb.tt', File.join('app', 'services', class_path, "#{file_name}.rb")
+  def copy_service_file
+    template 'service.rb.tt', File.join('app', 'services', class_path, "#{file_name}_service.rb")
   end
 
   hook_for :test_framework
