+++ lib/generators/service/templates/service.rb.tt
@@ -1,5 +1,5 @@
 <% module_namespacing do -%>
-class <%= class_name %>
+class <%= class_name %>Service
   def initialize(<%= attribute_names.join(", ") %>)
     <% attribute_names.each do |attribute| -%>
     @<%= attribute %> = <%= attribute %>
